// src/api/user/type.ts
// 登录接口需要携带参数ts类型
export  interface LoginFormData {
    username?: string
    password?: string
  }
 
  interface dataType{
    token?:string,
    message?:string
  }
  export interface ResponseData {
    code?: number
    message?: string
    ok?: boolean
  }
  
  export interface LoginResponseData extends ResponseData {
    data?: string
  }
  
  export interface userInfoResponseData extends ResponseData {
    data: {
      routes: string[]
      buttons: string[]
      roles: string[]
      name: string
      avatar: string
    }
  }